requirements.txt
--index-url https://pypi.org/simple
--extra-index-url https://download.pytorch.org/whl/cpu

streamlit==1.39.0
ultralytics==8.3.25
torch==2.9.1+cpu
opencv-python-headless==4.10.0.84
numpy==2.1.3
pillow==10.4.0